[IMP] mail, *: introduce public livechat bundle
Task-2990182
Task-2990191
