+++ addons/hr/static/src/models/messaging.js
@@ -2,7 +2,7 @@
 
 import { patchRecordMethods } from '@mail/model/model_core';
 // ensure that the model definition is loaded before the patch
-import '@mail/models/messaging';
+import '@mail/core_models/messaging';
 
 patchRecordMethods('Messaging', {
     /**
