+++ addons/im_livechat/static/src/models/messaging.js
@@ -1,10 +1,9 @@
 /** @odoo-module **/
 
-import { addFields, patchRecordMethods } from '@mail/model/model_core';
-import { many, one } from '@mail/model/model_field';
-import { clear } from '@mail/model/model_field_command';
+import { addFields } from '@mail/model/model_core';
+import { many } from '@mail/model/model_field';
 // ensure that the model definition is loaded before the patch
-import '@mail/models/messaging';
+import '@mail/core_models/messaging';
 
 addFields('Messaging', {
     /**
@@ -14,19 +13,4 @@ addFields('Messaging', {
         inverse: 'messagingAsPinnedLivechat',
         readonly: true,
     }),
-    publicLivechatGlobal: one('PublicLivechatGlobal', {
-        isCausal: true,
-    }),
-});
-
-patchRecordMethods('Messaging', {
-    /**
-     * @override
-     */
-    _computeNotificationHandler() {
-        if (this.publicLivechatGlobal) {
-            return clear();
-        }
-        return this._super();
-    },
 });
