+++ addons/im_livechat/static/src/public_models/chatbot.js
