+++ addons/im_livechat/static/src/public_models/chatbot_step.js
