+++ addons/im_livechat/static/src/public_models/livechat_button_view.js
