+++ addons/im_livechat/static/src/public_models/livechat_operator.js
