+++ addons/im_livechat/static/src/public_models/public_livechat_global_notification_handler.js
