+++ addons/im_livechat/static/src/public_models/public_livechat_view.js
