+++ addons/mail/static/src/core_models/device.js
