+++ addons/mail/static/src/core_models/locale.js
