+++ addons/mail/static/src/core_models/record.js
