+++ addons/website_livechat/static/src/public_models/chatbot.js
@@ -3,7 +3,7 @@
 import { addFields, patchRecordMethods } from '@mail/model/model_core';
 import { attr } from '@mail/model/model_field';
 // ensure that the model definition is loaded before the patch
-import '@im_livechat/models/chatbot';
+import '@im_livechat/public_models/chatbot';
 
 addFields('Chatbot', {
     isWebsiteLivechatTourFlow: attr({
