+++ addons/website_livechat/static/src/public_models/public_livechat_global.js
@@ -2,7 +2,7 @@
 
 import { patchRecordMethods } from '@mail/model/model_core';
 // ensure that the model definition is loaded before the patch
-import '@im_livechat/models/public_livechat_global';
+import '@im_livechat/public_models/public_livechat_global';
 
 import { set_cookie } from 'web.utils';
